JuliaCollider: now __buffer_size__, __unit_range_loop__ are Int32, faster

+++ stdlib/JuliaCollider/src/SCSynth.jl
@@ -4,7 +4,7 @@ export __SCSynth__
 
 struct __SCSynth__
     sampleRate::Float64
-    bufferSize::Int64
+    bufferSize::Int32
 end
 
 end
